Fixed Map Generation Algorithm
Fixed the map generation algorithm to automatically create new maps. Uses a seed so that the maps generated are always the same. (e.g. when you reach maze2, it will always have the same layout)

Added two new levels to the game. Adding new levels should be easier now with the algorithm.